requirements.txt
flask
streamlit
requests
web3
python-dotenv
twilio
plotly